refactor: atom, selector 파일을 도메인에 따라 분리
