+++ README.md
@@ -6,6 +6,7 @@ step 4 : Payments 페이지
 - [x] 해당 페이지에서 결제하기 버튼을 누르면 /order 요청을 보냄
 - [x] 결제 요청 후 성공 시, 장바구니 reset
 - [x] mockCoupons -> api의 "/coupons" GET 요청 데이터로 교체
+- [ ] 상품 추가 버튼 만들기
 
 step 3 : 적용된 쿠폰 계산 로직 구현
