+++ src/pages/checkout/components/CheckoutTotals.tsx
@@ -1,17 +1,16 @@
 import { useRecoilValue } from 'recoil';
+import {
+  orderAmountState,
+  totalAmountState,
+  totalShippingFeeState,
+} from '../../../store/orderStore';
+import { discountAmountState } from '../../../store/couponStore';
 import Divider from '../../../components/common/Divider/Divider';
 import NoticeLabel from '../../../components/common/NoticeLabel/NoticeLabel';
 import SubtitleSpaceBetween from '../../../components/common/SubtitleSpaceBetween/SubtitleSpaceBetween';
 import { NOTICE_MESSAGE } from '../../../constants/messages';
-
-import styles from '../Checkout.module.css';
-import {
-  discountAmountState,
-  orderAmountState,
-  totalAmountState,
-  totalShippingFeeState,
-} from '../../../store/selectors';
 import formatKoreanCurrency from '../../../utils/formatKoreanCurrency';
+import styles from '../Checkout.module.css';
 
 export default function CheckoutTotals() {
   const orderAmount = useRecoilValue(orderAmountState);
