+++ src/pages/checkout/components/CouponModal/CouponModal.tsx
@@ -1,11 +1,11 @@
+import { useRecoilValue } from 'recoil';
 import { Modal } from '@jaymyong66/simple-modal';
+import { discountAmountState } from '../../../../store/couponStore';
 import NoticeLabel from '../../../../components/common/NoticeLabel/NoticeLabel';
 import CouponList from './CouponList';
 import { COUPON_POLICY } from '../../../../constants/policy';
-import styles from './CouponModal.module.css';
-import { useRecoilValue } from 'recoil';
-import { discountAmountState } from '../../../../store/selectors';
 import formatKoreanCurrency from '../../../../utils/formatKoreanCurrency';
+import styles from './CouponModal.module.css';
 
 interface Props {
   isOpen: boolean;
